repl: re-organize location of REPL common lib files
