+++ .vs/launch.vs.json
@@ -141,13 +141,6 @@
     {
       "type": "default",
       "project": "CMakeLists.txt",
-      "projectTarget": "goalc.exe (bin\\goalc.exe)",
-      "name": "REPL - Auto Listen",
-      "args": ["--user-auto", "--auto-lt"]
-    },
-    {
-      "type": "default",
-      "project": "CMakeLists.txt",
       "projectTarget": "decompiler.exe (bin\\decompiler.exe)",
       "name": "Decompiler - Jak 1",
       "args": [
