+++ game/tools/subtitles/subtitle_editor.h
@@ -3,7 +3,7 @@
 #include <optional>
 #include <string_view>
 
-#include "common/nrepl/ReplClient.h"
+#include "common/repl/nrepl/ReplClient.h"
 #include "common/serialization/subtitles/subtitles_ser.h"
 
 #include "third-party/imgui/imgui.h"
